+++ lib/measures/SimplifyGeometryToSlicedBar/resources/os_lib_helper_methods.rb
@@ -353,10 +353,8 @@ module OsLib_HelperMethods
     objectArray.each do |object|
       object_LCCs = object.lifeCycleCosts
       object_LCCs.each do |object_LCC|
-        if object_LCC.category == category
-          if onlyYearFromStartZero == false || object_LCC.yearsFromStart == 0
-            counter += object_LCC.totalCost
-          end
+        if object_LCC.category == category && (onlyYearFromStartZero == false || object_LCC.yearsFromStart == 0)
+          counter += object_LCC.totalCost
         end
       end
     end