+++ lib/measures/create_DOE_prototype_building/tests/create_DOE_prototype_building_test.rb
@@ -37,7 +37,7 @@ require 'openstudio'
 require 'openstudio/measure/ShowRunnerOutput'
 require 'fileutils'
 
-require_relative '../measure.rb'
+require_relative '../measure'
 require 'minitest/autorun'
 
 require 'json'