+++ lib/measures/create_bar_from_deer_building_type_ratios/measure.rb
@@ -41,9 +41,9 @@ require 'openstudio-standards'
 # load OpenStudio measure libraries from openstudio-extension gem
 require 'openstudio-extension'
 require 'openstudio/extension/core/os_lib_helper_methods'
-require 'openstudio/extension/core/os_lib_geometry.rb'
-require 'openstudio/extension/core/os_lib_model_generation.rb'
-require 'openstudio/extension/core/os_lib_model_simplification.rb'
+require 'openstudio/extension/core/os_lib_geometry'
+require 'openstudio/extension/core/os_lib_model_generation'
+require 'openstudio/extension/core/os_lib_model_simplification'
 
 # start the measure
 class CreateBarFromDEERBuildingTypeRatios < OpenStudio::Measure::ModelMeasure