+++ lib/measures/create_typical_building_from_model/measure.rb
@@ -38,7 +38,7 @@ require 'openstudio-standards'
 # load OpenStudio measure libraries from openstudio-extension gem
 require 'openstudio-extension'
 require 'openstudio/extension/core/os_lib_helper_methods'
-require 'openstudio/extension/core/os_lib_model_generation.rb'
+require 'openstudio/extension/core/os_lib_model_generation'
 
 require_relative 'resources/Model.hvac' # DLM: should this be in openstudio-standards? dfg some tests fail without it