+++ lib/measures/create_typical_deer_building_from_model/tests/measure_test_hvac.rb
@@ -36,7 +36,7 @@
 require 'openstudio'
 require 'openstudio/measure/ShowRunnerOutput'
 require 'minitest/autorun'
-require_relative '../measure.rb'
+require_relative '../measure'
 require 'fileutils'
 
 class AddHVACSystemTest < MiniTest::Test