+++ lib/measures/merge_floorspace_js_with_model/tests/merge_floorspace_js_with_model_test.rb
@@ -38,7 +38,7 @@
 require 'openstudio'
 require 'openstudio/measure/ShowRunnerOutput'
 require 'minitest/autorun'
-require_relative '../measure.rb'
+require_relative '../measure'
 require 'fileutils'
 
 class MergeFloorspaceJsWithModelTest < Minitest::Test
@@ -51,7 +51,7 @@ class MergeFloorspaceJsWithModelTest < Minitest::Test
     # runner = OpenStudio::Measure::OSRunner.new(osw)
 
     # create an instance of a runner with OSW
-    osw_path = OpenStudio::Path.new(File.dirname(__FILE__) + '/measure_test.osw')
+    osw_path = OpenStudio::Path.new("#{File.dirname(__FILE__)}/measure_test.osw")
     osw = OpenStudio::WorkflowJSON.load(osw_path).get
     runner = OpenStudio::Measure::OSRunner.new(osw)